+++ src/main/java/com/service/Join.java
@@ -5,9 +5,9 @@ public class Join {
     public int id;
     public int originIdClient;
     public int destinyIdClient;
-    public int ratio;
+    public double ratio;
 
-    public Join(int id, int originIdClient, int destinyIdClient, int ratio) {
+    public Join(int id, int originIdClient, int destinyIdClient, double ratio) {
         this.id = id;
         this.originIdClient = originIdClient;
         this.destinyIdClient = destinyIdClient;
